[WIP] Cap'n'proto PoC
MVP is a multithreaded unit test with split-brain analogue SoloCoinJoin.

After that, there should be a program that can start tor daemon and
expose a capnproto RPC on STDIO, able to perform coinjoins on regtest.

The WabiSabi code should split out into a separate project that depends
on capnp to reduce a lot of boilerplate, this will allow implementing
the capnproto interfaces with the actual classes, and to provide
conversions from NBitcoin's types to the corresponding cap'n'proto
domain classes.

A major pain point is mutable properties of SpendableSmartCoin now that
it must be async. The client should just register a notification
callback for updates instead.

capnp.cs are files currently compiled manually to avoid issues with
codegen (needs deps from chocolatey).